fixed variable type in definition of bein hashmashos

+++ zmanim/zmanim.go
@@ -233,7 +233,7 @@ func (z *Zmanim) Tzeit(angle float64) time.Time {
 // One opinion on how to calculate this time is that it is 13.5 minutes before tzies 7.083
 func (z *Zmanim) BeinHashmashos() time.Time {
 	tzeis:=z.Tzeit(Tzeit3MediumStars)
-	return tzeis.Add(-time.Duration(13.5) * time.Minute)
+	return tzeis.Add(-time.Duration(13.5 * float64(time.Minute)))
 }
 
 func (z *Zmanim) riseSetOffset(t time.Time, offset int, roundTime bool) time.Time {
